requirements.txt
libsass==0.21.0
markdown==3.4.1
pillow==9.3.0
pystache==0.6.0
simple-http-server==0.18.3
